+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/amazon-bedrock
 
+## 0.0.6
+
+### Patch Changes
+
+- 42b11b8e: fix (provider/aws-bedrock): pass tool parameters for object generation without stringify
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [b7290943]
+  - @ai-sdk/provider@0.0.12
+  - @ai-sdk/provider-utils@1.0.2
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [d481729f]
+  - @ai-sdk/provider-utils@1.0.1
+
 ## 0.0.3
 
 ### Patch Changes