+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/google-vertex
 
+## 0.0.15
+
+### Patch Changes
+
+- bb584330: feat (provider/google-vertex): use systemInstruction content parts
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [b7290943]
+  - @ai-sdk/provider@0.0.12
+  - @ai-sdk/provider-utils@1.0.2
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [d481729f]
+  - @ai-sdk/provider-utils@1.0.1
+
 ## 0.0.12
 
 ### Patch Changes