+++ packages/react/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @ai-sdk/react
 
+## 0.0.20
+
+### Patch Changes
+
+- 56bbc2a7: feat (ai/ui): set body and headers directly on options for handleSubmit and append
+- Updated dependencies [56bbc2a7]
+  - @ai-sdk/ui-utils@0.0.12
+
+## 0.0.19
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@1.0.2
+- @ai-sdk/ui-utils@0.0.11
+
+## 0.0.18
+
+### Patch Changes
+
+- 70d18003: add setThreadId helper to switch between threads for useAssistant
+
+## 0.0.17
+
+### Patch Changes
+
+- 6a11cfaa: feat (ai/ui): add onError handler to useObject
+- 3db90c3d: allow empty handleSubmit submissions for useChat
+- Updated dependencies [d481729f]
+  - @ai-sdk/provider-utils@1.0.1
+  - @ai-sdk/ui-utils@0.0.10
+
 ## 0.0.16
 
 ### Patch Changes