+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ai-sdk/vue
 
+## 0.0.15
+
+### Patch Changes
+
+- dd0d854e: feat (ai/vue): add useAssistant
+
+## 0.0.14
+
+### Patch Changes
+
+- 56bbc2a7: feat (ai/ui): set body and headers directly on options for handleSubmit and append
+- Updated dependencies [56bbc2a7]
+  - @ai-sdk/ui-utils@0.0.12
+
+## 0.0.13
+
+### Patch Changes
+
+- @ai-sdk/ui-utils@0.0.11
+
+## 0.0.12
+
+### Patch Changes
+
+- 3db90c3d: allow empty handleSubmit submissions for useChat
+  - @ai-sdk/ui-utils@0.0.10
+
 ## 0.0.11
 
 ### Patch Changes